CompositeAnimationWidget revised, removes animation flipping logic.
